Build: Support Windows/MSVC using NMake
This adds a set of NMake Makefiles, with a set of support files that
is used to generate the full index.theme file, and to copy the icon/
cursor files in Windows, under Visual Studio builds.  Note that since
UNIX-style symlinks are not supported under Windows, we generate a
.bat script during 'make dist' so that we can copy the real files that
the symlinks point to, to replace the symlinks on Windows.