anicursorgen: change default shadow color
0x80808080 was a mistake, it produces good-looking shadow on white
background, but doesn't work as intended on darker backgrounds.
Instead of making the shadow gray and 50% opaque,
make it completely black, but only 25% opaque. This way it still
looks appropriately-grey on white background, but quickly fades
to black as the background gets darker. Or, to put it another way,
it doesn't lighten dark backrounds.

https://bugzilla.gnome.org/show_bug.cgi?id=763034