Add directories containing only prerendered symbolic icons to index.theme
If some directory contains no rendered icons or svgs, then the code
we use to detect the list of directories to put in the index.theme file
will not add them.

These dirs will still have files in them (generated by the
gtk-encode-symbolic-svg), but since they are not in the index.theme
they will not be used.

We fix this by generating a .placeholder file in such directories
so that the index.theme generation works.