meson: Create duplicate cursors via symlinks at install time
In !49 (Commit 7a5dc9d5), the duplicate cursors (aliases for cursor
names) were switched from being symlinks to being copies to avoid
problems with using symlinks with meson's `install_subdir` function.
However, copying the cursors needlessly inflates the size of the
distribution tarball and distribution packages.

The cursors are only installed on unix-like systems, so the symlinks
themselves are not a problem.

To solve this problem, I have deleted the duplicate cursor files, and
set up meson to create the required symlinks at install time. The list
of symlinks to create is derived from the list of copied symlinks from
the `make.sh` file (it has been updated to no longer create the copies).